Plot NimbledroidGraph in its own view
